refactor(WireGuard): Move client traffic restriction to WireGuard class
